# aprendi-api/cmd/aprendi-api/docs/swagger.yaml
definitions:
  repos.CourseModel:
    properties:
      course_name:
        description: Name of the course
        type: string
      description:
        description: Description of the course content
        type: string
      id:
        description: Unique identifier for the course
        type: string
      org_id:
        description: Identifier for the organization
        type: string
      section:
        description: Section or batch of the course
        type: string
    type: object
  repos.CourseRequest:
    properties:
      course_id:
        type: string
      priority:
        type: integer
    type: object
  repos.EnrollmentRequest:
    properties:
      courses:
        items:
          $ref: '#/definitions/repos.CourseRequest'
        type: array
      org_id:
        type: string
      student_id:
        type: string
      term_id:
        type: string
    type: object
  repos.OrganizationModel:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
info:
  contact: {}
paths:
  /courses:
    post:
      consumes:
      - application/json
      description: Save a new course
      parameters:
      - description: Course Model
        in: body
        name: course
        required: true
        schema:
          $ref: '#/definitions/repos.CourseModel'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/repos.CourseModel'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a course
      tags:
      - courses
  /courses/{orgID}:
    get:
      consumes:
      - application/json
      description: Retrieve all courses
      parameters:
      - description: Organization ID
        in: path
        name: orgID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/repos.CourseModel'
            type: array
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all courses under an orgID
      tags:
      - courses
  /courses/{orgID}/{courseID}:
    get:
      consumes:
      - application/json
      description: Retrieve course details
      parameters:
      - description: Organization ID
        in: path
        name: orgID
        required: true
        type: string
      - description: Course ID
        in: path
        name: courseID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repos.CourseModel'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a course by orgID and courseID
      tags:
      - courses
  /enroll:
    post:
      consumes:
      - application/json
      description: enroll a student to a course
      parameters:
      - description: Enrollment Request
        in: body
        name: enrollment
        required: true
        schema:
          $ref: '#/definitions/repos.EnrollmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Enroll a student
      tags:
      - students
  /example/helloworld:
    get:
      consumes:
      - application/json
      description: do ping
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: ping example
      tags:
      - example
  /organizations:
    get:
      consumes:
      - application/json
      description: Retrieve all organizations
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/repos.OrganizationModel'
            type: array
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all organizations
      tags:
      - organizations
    post:
      consumes:
      - application/json
      description: Save a new organization
      parameters:
      - description: Organization Model
        in: body
        name: organization
        required: true
        schema:
          $ref: '#/definitions/repos.OrganizationModel'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/repos.OrganizationModel'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create an organization
      tags:
      - organizations
  /organizations/{orgID}:
    get:
      consumes:
      - application/json
      description: Retrieve organization details
      parameters:
      - description: Organization ID
        in: path
        name: orgID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repos.OrganizationModel'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get an organization by its ID
      tags:
      - organizations
swagger: "2.0"